dom/indexedDB/test/test_remove_index.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Property Test</title>

  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript;version=1.7">
    function testSteps()
    {
      const nsIIDBObjectStore = Components.interfaces.nsIIDBObjectStore;

      const name = window.location.pathname;
      const description = "My Test Database";
      const indexName = "My Test Index";

      let request = moz_indexedDB.open(name, description);
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      let event = yield;

      let db = event.result;
      is(db.objectStoreNames.length, 0, "Correct objectStoreNames list");

      request = db.setVersion("1");
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      event = yield;

      let objectStore = db.createObjectStore("test store", "foo");
      is(db.objectStoreNames.length, 1, "Correct objectStoreNames list");
      is(db.objectStoreNames.item(0), objectStore.name, "Correct name");

      is(objectStore.indexNames.length, 0, "Correct indexNames list");

      objectStore.createIndex(indexName, "foo");

      is(objectStore.indexNames.length, 1, "Correct indexNames list");
      is(objectStore.indexNames.item(0), indexName, "Correct name");

      objectStore.removeIndex(indexName);

      is(event.result, undefined, "Correct result");
      is(objectStore.indexNames.length, 0, "Correct indexNames list");

      finishTest();
      yield;
    }
  </script>
  <script type="text/javascript;version=1.7" src="helpers.js"></script>

</head>

<body onload="runTest();"></body>

</html>
